add foreign key to romm and cinema

+++ Modules/Seat/Database/Migrations/2024_05_08_133231_create_seat_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->boolean('status')->default(false);
             $table->string('type');
             $table->unique(['row', 'column', 'room_id']);
-            $table->unsignedInteger('room_id');
+            $table->unsignedBigInteger('room_id');
             $table->timestamps();
         });
     }
